refactor github_api module
